Extract ComponentDescriptor.Factory to its own file.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=230565160
